uav/requirements.txt
future==1.0.0
loguru==0.7.3
lxml==5.3.0
pymavlink==2.4.42
python-dotenv==1.0.1
